🐛 Catch Reddit Urls with no comment id and invalid url
Previously, the API allowed urls to entire posts and not specific
comments since the regular expression used * as opposed to +.
In the event of no match ([][0] -> undefined), the redditUrl property
would also have been left undefined.
Now, the API will return an error if no url regexp match was found, as
well as when the actual link fails to validate.